+++ Changelog.md
@@ -3,6 +3,7 @@
 _Changes:_
 
 - [X] PDO-8355 Cronjob delete Job and PersistentVolumeClaim resources for PingDirectory backups
+- [X] PDO-8362 OpenSearch: Add app_timestamp field
 - [X] PDO-8363 OpenSearch: Add the Fluent Bit ingestion time field